refactor: removed unused examples

+++ examples/i_mesh_change.cpp
@@ -140,7 +140,7 @@ int main()
 	}
 	
 	float initialTime = glfwGetTime();
-	float lastTime = initialTime;
+	double lastTime = initialTime;
 	bool toSphereMesh = true;
 
 	// Main while loop
@@ -169,8 +169,8 @@ int main()
 			initialTime = lastTime;
 		}
 
-		float currentTime = glfwGetTime();
-		float dt = currentTime - lastTime;
+		double currentTime = glfwGetTime();
+		double dt = currentTime - lastTime;
 		lastTime = currentTime;
 		
 		updateFromInputs(window, dt, &camera);
